+++ RayTracerChallengev3/src/Patterns/RingPattern.cpp
@@ -6,7 +6,7 @@ namespace RayTracer
 {
 	Color RingPattern::patternAt(const Point4& p) const
 	{
-		int val = (int)std::floor(std::sqrtf(p.x * p.x + p.z * p.z));
+		int val = (int)std::floor(std::sqrt(p.x * p.x + p.z * p.z));
 		if (val % 2 == 0)
 		{
 			return color1;
